AS7-1361 Move Modules Dependencies 'enrichment' to Protocol-JMX-AS7.
The JMX-AS7 Protocol needs to enrich the Deployed archive with Dependencies on the Arquillian Service. This enrichment needs to be a part of the Protocol and not Container to avoid the same enrcihment leak into other Protocols e.g. Servlet. Adding these Dependencies with out deploying the Arquillian Service cause deployment exception.
